refactor: components categorized
