+++ packages/front/package.json
@@ -9,9 +9,11 @@
   },
   "dependencies": {
     "@mdi/font": "^6.4.95",
+    "@types/leaflet": "^1.7.5",
     "element-plus": "1.1.0-beta.20",
     "github-markdown-css": "^5.0.0",
     "html-loader": "^1.3.2",
+    "leaflet": "^1.7.1",
     "markdown-loader": "^6.0.0",
     "vue": "^3.2.20",
     "vue-router": "^4.0.12"
