+++ packages/front/src/components/layout/Footer.vue
