+++ packages/front/src/components/slider/SliderPlayButton.vue
